feat: initialize cache dirs in harvest and check in process
We want to abort execution if process step cannot find any cache
from harvest (which means the step did not run) and present an error to the user.

If the harvesters itself did not run during the step (because disabled), this should not be
a problem, as in the future we will only look for data from enabled harvesters.