docs(init,slim_click): move docstring for ColorLogFormatter to class level
When used on the init method, the original docstrings for the base class are used. They provoke docs build failures, as they contain line breaks which Sphinx doesn't like (they are seen as blockquotes).

As we treat all warnings as errors for docs builds, this breaks the workflows for our docs. Moving the docstring from _init to the class itself resolved the problem.