chore: fix pyproject's authors and maintainers metadata
We interpret PEP-621 the following way:

- "Authors" within the Python package metadata will be the legal entities holding the intellectual property, as this list is also used to express copyright statements.
- "Maintainers" will enlist the people involved for a longer period of time, doing contributions on a regular basis and usually equipped with write access to the repository.

This is different from the definition of authors for a scholarly publication, which is managed in CITATION.cff where order matters, too.

Both maintainers and authors will be ordered in alphabetical order without any intent of prioritization or importance.

There are no rules yet when a maintainer should be dropped of the list or another legal entity should be added to authors.

See also: https://peps.python.org/pep-0621/