docs: update release process for CFF version and release branch
As discussed with @sdruskat.

We will not be updating the CFF to a new dev version in the last step. The repo default branch shall hold the latest released information for a project, not a dev snap.

This also means that any project using HERMES probably should do any changes to CITATION.cff metadata as late as possible.